requirements.txt
fastapi
uvicorn
jinja2
python-multipart
paho-mqtt